Upgrade config file to match new schemas.
When a new schema is introduced, now config files (base config and profiles)
can be upgraded to meet the new schema. The schemaVersion (in
configSpec.py) should be incremented, a new function should be added
to the profileUpgradeSteps.py file to make the changes required to the
file.

After all upgrade steps are made the file is validated, and any errors
will be in the log.